--story=115744045 OPdata增加gpu卡库存视图

+++ src/backend/ci/ext/tencent/dispatch/biz-dispatch-kubernetes-tencent/src/main/kotlin/com/tencent/devops/dispatch/kubernetes/startcloud/service/StartCloudInterfaceService.kt
@@ -252,7 +252,10 @@ class StartCloudInterfaceService @Autowired constructor(
             }
         }
         logger.debug("get all vm resource|resourceList|{}", resList)
-        windowsGpuResourceDao.deleteVmResource(dslContext)
-        windowsGpuResourceDao.insertVmResource(dslContext, resList)
+        if (resList.isNotEmpty()) {
+            windowsGpuResourceDao.deleteVmResource(dslContext)
+            windowsGpuResourceDao.insertVmResource(dslContext, resList)
+        }
+
     }
 }
